resolve fixes and suggestions by edgar
