+++ app/middlewares/schemaValidate.js
@@ -1,0 +1,14 @@
+const { validationResult } = require('express-validator');
+
+exports.schemaValidate = (req, res, next) => {
+  const errors = validationResult(req);
+  if (errors.isEmpty()) {
+    return next();
+  }
+  const extractedErrors = [];
+  errors.array().map(err => extractedErrors.push({ [err.param]: err.msg }));
+
+  return res.status(400).json({
+    errors: errors.array()
+  });
+};
