+++ app/schemas/users.js
@@ -1,4 +1,4 @@
-const { body, validationResult } = require('express-validator');
+const { body } = require('express-validator');
 const { existUserMail } = require('../services/users');
 const config = require('../../config');
 
@@ -47,16 +47,3 @@ exports.signUpSchema = () => [
     .isLength({ min: 8 })
     .withMessage('Password should be at least 8 chars long')
 ];
-
-exports.validate = (req, res, next) => {
-  const errors = validationResult(req);
-  if (errors.isEmpty()) {
-    return next();
-  }
-  const extractedErrors = [];
-  errors.array().map(err => extractedErrors.push({ [err.param]: err.msg }));
-
-  return res.status(422).json({
-    errors: extractedErrors
-  });
-};
