+++ app/services/random_quotes.js
@@ -4,7 +4,7 @@ const config = require('../../config');
 
 const API_URL = config.common.external.quoteApiUrl;
 
-const getQuote = async (number = 'random', type = '') => {
+exports.getQuote = async (number = 'random', type = '') => {
   try {
     const typePrefix = type ? '/' : '';
     const url = `${API_URL}/${number}${typePrefix}${type}`;
@@ -15,5 +15,3 @@ const getQuote = async (number = 'random', type = '') => {
     return new Error(`getQuote Error => ${error.message}`);
   }
 };
-
-module.exports = { getQuote };
